+++ internal/app/wwctl/node/delete/main.go
@@ -18,7 +18,7 @@ func CobraRunE(cmd *cobra.Command, args []string) (err error) {
 	}
 
 	if !SetYes {
-		var nodeList []node.NodeConf
+		var nodeList []node.Node
 		// The checks run twice in the prompt case.
 		// Avoiding putting in a blocking prompt in an API.
 		nodeList, err = apiNode.NodeDeleteParameterCheck(&ndp, false)