+++ internal/app/wwctl/power/soft/root.go
@@ -5,13 +5,20 @@ import (
 	"github.com/warewulf/warewulf/internal/pkg/node"
 )
 
-var (
-	powerCmd = &cobra.Command{
+type variables struct {
+	Showcmd bool
+	Fanout  int
+}
+
+// GetRootCommand returns the root cobra.Command for the application.
+func GetCommand() *cobra.Command {
+	vars := variables{}
+	powerCmd := &cobra.Command{
 		DisableFlagsInUseLine: true,
 		Use:                   "soft",
 		Short:                 "Gracefully shuts down the given node(s)",
 		Long:                  "This command uses the operating system to shut down the set of nodes specified by PATTERN.",
-		RunE:                  CobraRunE,
+		RunE:                  CobraRunE(&vars),
 		ValidArgsFunction: func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
 			if len(args) != 0 {
 				return nil, cobra.ShellCompDirectiveNoFileComp
@@ -26,9 +33,7 @@ var (
 			return node_names, cobra.ShellCompDirectiveNoFileComp
 		},
 	}
-)
-
-// GetRootCommand returns the root cobra.Command for the application.
-func GetCommand() *cobra.Command {
+	powerCmd.PersistentFlags().BoolVarP(&vars.Showcmd, "show", "s", false, "only show command which will be executed")
+	powerCmd.PersistentFlags().IntVar(&vars.Fanout, "fanout", 50, "how many command should be executed in parallel")
 	return powerCmd
 }